+++ package.json
@@ -10,6 +10,7 @@
     "apollo-cache-inmemory": "^1.6.3",
     "apollo-client": "^2.6.4",
     "apollo-link-http": "^1.5.16",
+    "axios": "^0.24.0",
     "graphql": "^15.7.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
