+++ src/Components/Dashboard.css
@@ -15,8 +15,8 @@ td{
 
 
 #csv-form{
-    margin-top: -750px;
-    margin-left: 750px;
+    margin-top: 120px;
+    margin-left: 550px;
 
 }
 
@@ -24,7 +24,7 @@ table{
     min-width: 900px;
 }
 
-.csv-button{
+.csv-button,.upload-button{
     background-color: #090a0a;
     color: #eeeeee;
     border:  2px solid #10fbd0;
@@ -51,4 +51,15 @@ table{
     margin-left: 10px;
 
 
+}
+
+#csv-form>h2{
+
+    margin:20px;
+    font-size:25px;
+}
+
+
+.upload-button{
+    margin:20px;
 }
